refactor: separate non-SQL models to separate file
