+++ processing/app/models.py
@@ -3,7 +3,6 @@ from datetime import datetime
 from enum import StrEnum
 from typing import Optional
 
-from pydantic import BaseModel
 from pydantic import UUID4
 from sqlmodel import Column
 from sqlmodel import Field
@@ -45,48 +44,3 @@ class SurveyAnswer(SQLModel, table=True):
 
     survey_id: UUID4 = Field(foreign_key="survey.id")
     survey: Optional["Survey"] = Relationship(back_populates="answers")
-
-
-class AnalysisJobResponse(BaseModel):
-    id: UUID4
-    survey_id: UUID4
-    created_at: datetime
-    status: TaskStatus
-
-
-class AnalysisTaskResult(BaseModel):
-    result: str | None
-    survey_id: UUID4
-    status: TaskStatus
-
-
-class ProcessSurveyRequest(BaseModel):
-    survey_id: UUID4
-    question: str
-    answers: list[str]
-
-
-class SurveyStatusReponse(BaseModel):
-    id: UUID4
-    survey_id: str
-    created_at: datetime
-    status: TaskStatus
-
-
-class ProcessSurveyResponse(BaseModel):
-    llm_response: str
-
-
-class SurveyPointsSentimentBucket(BaseModel):
-    """Sentiment bucket for survey points"""
-
-    positive: list[str] = Field(description="List of positive points")
-    negative: list[str] = Field(description="List of negative points")
-
-
-class SurveyPoints(BaseModel):
-    """Data that we want collect from llm as response"""
-
-    frequent: SurveyPointsSentimentBucket = Field(description="List of frequent points")
-    moderate: SurveyPointsSentimentBucket = Field(description="List of moderate points")
-    occasional: SurveyPointsSentimentBucket = Field(description="List of occasional points")
