+++ processing/app/tests/test_analysis.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 from uuid import uuid4
 
 import pytest
-from app.models import ProcessSurveyRequest
+from app.schemas import ProcessSurveyRequest
 from app.services.analysis_service import AnalysisService
